pyproject.toml
[tool.poetry]
name = "youtube downloader"
version = "0.1.0"
description = ""
authors = ["amir <amirnurani2001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytube = "^15.0.0"
pyyaml = "^6.0.2"
pydantic-settings = "^2.7.0"
yt-dlp = "^2024.12.23"
ffmpeg = "^1.4"
ffprobe = "^0.5"
httpx = "^0.28.1"
python-telegram-bot = "^21.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
